Consolidate extensions
